feat(cli): Config refactor and download support started
